fixing repo structure
