+++ .gitignore
@@ -1,0 +1,1 @@
+software/node_modules
