--- software/1.0/webapp/1.0/pages/header.html
@@ -1,8 +1,0 @@
-<template>
-	<div class="panel panel-warning" id="panel_connectionStatus">
-		<div class="panel-heading">
-			<div style="display: inline-block; padding-left: 35%;" class="panel-title text-center">Estado: <strong id="connection_status" class="text-danger">Desconectado</strong> </div>
-			<button type="button" class="btn btn-xs btn-default" id="btn_connect">Conectar</button>
-		</div>
-	</div>
-</template>
