--- software/1.0/webapp/1.0/pages/info.html
@@ -1,34 +1,0 @@
-<template>
-<div class="panel-body">
-	<div id="info-rotation">
-		<h5 class="text-center">Rotacion</h5>
-		<label class="btn-block">x: <span class="label label-success">0</span></label>
-		<label class="btn-block">y: <span class="label label-success">0</span></label>
-		<label class="btn-block">z: <span class="label label-success">0</span></label>
-	</div>
-
-	<div id="info-motors">
-		<h5 class="text-center">Motores</h5>
-
-		<span id="motor0">Izquierdo: <label>0</label>%</span>
-		<div class="progress">
-			<div class="progress-bar progress-bar-info" style="width: 0%;" role="progressbar"></div>
-		</div>
-
-		<span id="motor2">Derecho: <label>0</label>%</span>
-		<div class="progress">
-			<div class="progress-bar progress-bar-success" style="width: 0%;" role="progressbar"></div>
-		</div>
-
-		<span id="motor1">Frontal: <label>0</label>%</span>
-		<div class="progress">
-			<div class="progress-bar progress-bar-warning" style="width: 0%;" role="progressbar"></div>
-		</div>
-
-		<span id="motor3">Trasero: <label>0</label>%</span>
-		<div class="progress">
-			<div class="progress-bar progress-bar-danger" style="width: 0%;" role="progressbar"></div>
-		</div>
-
-	</div>
-</div></template>
