+++ software/nop_modules/PID.js
@@ -12,7 +12,9 @@ function meConstructor(type, setPoint, kp, ki, kd) {
 	this.lastError = 0;
 };
 
-function compute(angleData) {
+function compute(angleData, setPoint) {
+	this.setPoint = setPoint;
+	console.log("SP: ", this.setPoint);
 	var input;
 	switch(this.type.toUpperCase()) {
 		case "ROLL":
@@ -25,6 +27,10 @@ function compute(angleData) {
 			input = angleData.z;
 		break;
 	}
+	if (!input || isNaN(input)) {
+		console.log("Cannot compute, input is NULL");
+		return 0;
+	}
 	var error = this.setPoint - input;
 	this.errorArea += error;
 	var derivativeError = (error - this.lastError);
@@ -33,11 +39,13 @@ function compute(angleData) {
 	var I = this.ki * this.errorArea;
 	var D = this.kd * derivativeError;
 
+	I = 0; //TODO: Fix integral problem
+
 	var val = P + I + D;
 	this.lastError = error;
 
-	val = val > 255 ? 255 : val;
-	//val = val < 0 ? 0 : val;
+	val = val > 30 ? 30 : val;
+	val = val < -30 ? -30 : val;
 
 	return Math.round(val);
 };
